REACT defaultProps
Default props in React are values that are used when a component is rendered without a specific prop being passed in. They provide a fallback value to prevent errors or unexpected behavior.  You can define default props using the `defaultProps` property on a component. For example, you could define a default prop for a `name` prop on a `User` component like this:

```javascript
User.defaultProps = {
  name: 'Guest'
};
```

This would ensure that the `name` prop always has a value, even if it's not passed in from the parent component.